Refactor maven resolver: split ProjectBuilder, enforce one-class-per-file, remove unused constructors
