+++ detectable/src/main/java/com/blackduck/integration/detectable/detectables/maven/resolver/MavenResolverOptions.java
@@ -76,88 +76,6 @@ public class MavenResolverOptions {
     }
 
     /**
-     * Constructs MavenResolverOptions with external repositories and individual proxy fields (backward compatibility).
-     *
-     * <p>This constructor is provided for backward compatibility with existing callers.
-     * New code should use the constructor that accepts {@link MavenProxyConfig}.
-     *
-     * @param externalRepositories List of external Maven repository URLs to use during resolution.
-     * @param proxyHost            Proxy hostname or IP (no scheme). May be null if no proxy is configured.
-     * @param proxyPort            Proxy port number (0 if not configured).
-     * @param proxyUsername        Proxy authentication username. May be null.
-     * @param proxyPassword        Proxy authentication password. May be null.
-     * @param proxyIgnoredHosts    Host patterns that should bypass the proxy. Never null.
-     * @param mirrorConfigurations  List of mirror configurations for corporate repository managers.
-     */
-    public MavenResolverOptions(
-        List<String> externalRepositories,
-        @Nullable String proxyHost,
-        int proxyPort,
-        @Nullable String proxyUsername,
-        @Nullable String proxyPassword,
-        List<String> proxyIgnoredHosts,
-        List<MavenMirrorConfig> mirrorConfigurations
-    ) {
-        this.externalRepositories = externalRepositories != null ? externalRepositories : Collections.emptyList();
-
-        // Convert individual proxy fields to MavenProxyConfig if proxy is configured
-        if (proxyHost != null && !proxyHost.trim().isEmpty() && proxyPort > 0) {
-            this.proxyConfig = new MavenProxyConfig(
-                proxyHost,
-                proxyPort,
-                proxyUsername,
-                proxyPassword,
-                proxyIgnoredHosts != null ? proxyIgnoredHosts : Collections.emptyList()
-            );
-        } else {
-            this.proxyConfig = null;
-        }
-
-        this.mirrorConfigurations = mirrorConfigurations != null ? mirrorConfigurations : Collections.emptyList();
-        this.includeTestScope = true; // Default to true for backward compatibility
-    }
-
-    /**
-     * Constructs MavenResolverOptions with external repositories and proxy configuration (no mirrors).
-     *
-     * <p>This constructor is provided for backward compatibility with existing callers.
-     *
-     * @param externalRepositories List of external Maven repository URLs to use during resolution.
-     *                             These are used alongside repositories declared in pom.xml.
-     * @param proxyHost            Proxy hostname or IP (no scheme). May be null if no proxy is configured.
-     * @param proxyPort            Proxy port number (0 if not configured).
-     * @param proxyUsername        Proxy authentication username. May be null.
-     * @param proxyPassword        Proxy authentication password. May be null.
-     * @param proxyIgnoredHosts    Host patterns that should bypass the proxy. Never null.
-     */
-    public MavenResolverOptions(
-        List<String> externalRepositories,
-        @Nullable String proxyHost,
-        int proxyPort,
-        @Nullable String proxyUsername,
-        @Nullable String proxyPassword,
-        List<String> proxyIgnoredHosts
-    ) {
-        this.externalRepositories = externalRepositories != null ? externalRepositories : Collections.emptyList();
-
-        // Convert individual proxy fields to MavenProxyConfig if proxy is configured
-        if (proxyHost != null && !proxyHost.trim().isEmpty() && proxyPort > 0) {
-            this.proxyConfig = new MavenProxyConfig(
-                proxyHost,
-                proxyPort,
-                proxyUsername,
-                proxyPassword,
-                proxyIgnoredHosts != null ? proxyIgnoredHosts : Collections.emptyList()
-            );
-        } else {
-            this.proxyConfig = null;
-        }
-
-        this.mirrorConfigurations = Collections.emptyList();
-        this.includeTestScope = true; // Default to true for backward compatibility
-    }
-
-    /**
      * Returns the list of external Maven repository URLs.
      *
      * @return List of repository URLs, never null
