+++ detectable/src/main/java/com/blackduck/integration/detectable/detectables/maven/resolver/ProjectBuilder.java
@@ -1,7 +1,6 @@
 package com.blackduck.integration.detectable.detectables.maven.resolver;
 
 import com.blackduck.integration.detectable.detectables.maven.resolver.model.*;
-import org.apache.commons.text.StringSubstitutor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -9,7 +8,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -27,9 +25,11 @@ import java.util.stream.Collectors;
  * Responsibilities:
  * - Parse and interpolate POM files using {@link PomParser} and {@link PropertiesResolverProvider}.
  * - Download parent and BOM POMs via {@link MavenDownloader} when local parent POMs are not available.
- * - Merge properties, repositories, dependency management, and dependencies from parent/BOM models.
- * - Apply dependency management to resolve missing versions/scopes and produce the final dependency list.
- * - Recursively process nested BOMs (BOMs that import other BOMs) with cycle detection.
+ * - Merge properties, repositories, dependency management, and dependencies from parent/BOM models
+ *   (delegated to {@link ModelMerger}).
+ * - Process imported BOMs (delegated to {@link BomProcessor}).
+ * - Apply dependency management to resolve missing versions/scopes and produce the final dependency
+ *   list (delegated to {@link DependencyConverter}).
  *
  * The implementation caches partially-built projects to avoid duplicate work and detects parent
  * cycles to prevent infinite recursion. Instances are not thread-safe due to internal mutable state
@@ -41,22 +41,24 @@ import java.util.stream.Collectors;
  */
 public class ProjectBuilder {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
-    
-    /**
-     * Maximum number of BOM processing iterations to prevent infinite loops from pathological POMs.
-     * This is a safety limit - in practice, most projects have far fewer nested BOMs.
-     */
-    private static final int MAX_BOM_ITERATIONS = 50;
-    
+
     private final PomParser pomParser;
     private final PropertiesResolverProvider propertiesResolverProvider;
     private final Map<String, PartialMavenProject> pomCache = new HashMap<>();
     private final Path downloadDir;
 
+    private final BomProcessor bomProcessor;
+    private final ModelMerger modelMerger;
+    private final DependencyConverter dependencyConverter;
+
     public ProjectBuilder(Path downloadDir) {
         this.pomParser = new PomParser();
         this.propertiesResolverProvider = new PropertiesResolverProvider(null, System::getenv);
         this.downloadDir = downloadDir;
+
+        this.bomProcessor = new BomProcessor(downloadDir, this, propertiesResolverProvider);
+        this.modelMerger = new ModelMerger();
+        this.dependencyConverter = new DependencyConverter(propertiesResolverProvider);
     }
 
     public MavenProject buildProject(File pomFile) throws Exception {
@@ -72,7 +74,7 @@ public class ProjectBuilder {
         return toCompleteMavenProject(canonicalPom.getCanonicalPath(), partialMavenProject);
     }
 
-    private PartialMavenProject internalBuildProject(File pomFile, Set<String> identifiedParents) throws Exception {
+    PartialMavenProject internalBuildProject(File pomFile, Set<String> identifiedParents) throws Exception {
         // Canonicalize the file to normalize path segments like '..' and symlinks so cache keys match
         File canonicalFile = pomFile;
         try {
@@ -99,7 +101,7 @@ public class ProjectBuilder {
                 String parentKey = buildGavKey(parentCoords);
                 if (identifiedParents.contains(parentKey)) {
                     logger.debug("Cycle detected for parent key '{}', treating file \"{}\" as direct super pom descendant.", parentKey, pomFilePath);
-                    return finalizeEffectiveModel(pomFilePath, pomFileInfo, null);
+                    return modelMerger.finalizeEffectiveModel(pomFilePath, pomFileInfo, null, pomCache);
                 }
                 identifiedParents.add(parentKey);
 
@@ -117,7 +119,7 @@ public class ProjectBuilder {
                         PartialMavenProject effectiveParentPom = internalBuildProject(expectedParentFile, identifiedParents);
                         propertiesResolverProvider.setParentProperties(effectiveParentPom.getProperties());
                         PartialMavenProject interpolatedPomInfo = pomParser.parsePomFile(pomFilePath, content, propertiesResolverProvider);
-                        return finalizeEffectiveModel(pomFilePath, interpolatedPomInfo, effectiveParentPom);
+                        return modelMerger.finalizeEffectiveModel(pomFilePath, interpolatedPomInfo, effectiveParentPom, pomCache);
                     } else {
                         logger.debug("No local parent POM found at expected path: {}", expectedParentPath);
                     }
@@ -144,345 +146,44 @@ public class ProjectBuilder {
 
                     propertiesResolverProvider.setParentProperties(effectiveParentPom.getProperties());
                     PartialMavenProject interpolatedPomInfo = pomParser.parsePomFile(pomFilePath, content, propertiesResolverProvider);
-                    return finalizeEffectiveModel(pomFilePath, interpolatedPomInfo, effectiveParentPom);
+                    return modelMerger.finalizeEffectiveModel(pomFilePath, interpolatedPomInfo, effectiveParentPom, pomCache);
                 } else {
                     logger.warn("Could not retrieve parent pom for file: \"{}\"", pomFilePath);
                 }
             }
         }
 
-        return finalizeEffectiveModel(pomFilePath, pomFileInfo, null);
+        return modelMerger.finalizeEffectiveModel(pomFilePath, pomFileInfo, null, pomCache);
     }
 
     private String buildGavKey(JavaCoordinates coords) {
         return coords.getGroupId() + ":" + coords.getArtifactId() + ":" + coords.getVersion();
     }
 
-    private void resolveDependencyProperties(PartialMavenProject project) {
-        // Create a complete map of properties for resolution.
-        // Build a combined properties map that includes parent properties so
-        // BOM coordinates that use parent/root properties (e.g. ${spring-cloud-dependencies.version})
-        // can be resolved during the first pass.
-        Map<String, String> allProps = new HashMap<>();
-        try {
-            Map<String, String> parentProps = propertiesResolverProvider.getParentProperties();
-            if (parentProps != null) {
-                allProps.putAll(parentProps);
-            }
-        } catch (Exception e) {
-            logger.debug("Failed to include parent properties in resolution map: {}", e.getMessage());
-        }
-        if (project.getProperties() != null) {
-            allProps.putAll(project.getProperties());
-        }
-
-        // Resolve properties for dependencies
-        for (PomXmlDependency dep : project.getDependencies()) {
-            dep.setGroupId(resolveProperties(dep.getGroupId(), allProps));
-            dep.setArtifactId(resolveProperties(dep.getArtifactId(), allProps));
-            dep.setVersion(resolveProperties(dep.getVersion(), allProps));
-        }
-        // Resolve properties for dependency management
-        for (PomXmlDependency dep : project.getDependencyManagement()) {
-            dep.setGroupId(resolveProperties(dep.getGroupId(), allProps));
-            dep.setArtifactId(resolveProperties(dep.getArtifactId(), allProps));
-            dep.setVersion(resolveProperties(dep.getVersion(), allProps));
-        }
-    }
-
-    private String resolveProperties(String value, Map<String, String> properties) {
-        if (value == null || !value.contains("${")) {
-            return value;
-        }
-        // Use StringSubstitutor for more robust property replacement.
-        StringSubstitutor sub = new StringSubstitutor(properties);
-        return sub.replace(value);
-    }
-
-    private List<JavaDependency> applyDependencyManagement(PartialMavenProject project) {
-        Map<String, PomXmlDependency> managedDependencies = project.getDependencyManagement().stream()
-            .collect(Collectors.toMap(dep -> dep.getGroupId() + ":" + dep.getArtifactId(), dep -> dep));
-
-        return project.getDependencies().stream()
-            .map(dep -> {
-                PomXmlDependency managed = managedDependencies.get(dep.getGroupId() + ":" + dep.getArtifactId());
-                if (managed != null) {
-                    if (isEmpty(dep.getVersion()) && !isEmpty(managed.getVersion())) {
-                        dep.setVersion(managed.getVersion());
-                    }
-                    if (isEmpty(dep.getScope()) && !isEmpty(managed.getScope())) {
-                        dep.setScope(managed.getScope());
-                    }
-                }
-                return convertPomXmlDependencyToJavaDependency(dep);
-            })
-            .collect(Collectors.toList());
-    }
-
-    private JavaDependency convertPomXmlDependencyToJavaDependency(PomXmlDependency pomDep) {
-        List<JavaDependencyExclusion> exclusions = new ArrayList<>();
-        if (pomDep.getExclusions() != null) {
-            exclusions = pomDep.getExclusions().stream()
-                .map(exclusion -> new JavaDependencyExclusion(exclusion.getGroupId(), exclusion.getArtifactId()))
-                .collect(Collectors.toList());
-        }
-
-        JavaCoordinates coordinates = new JavaCoordinates(pomDep.getGroupId(), pomDep.getArtifactId(), pomDep.getVersion(), pomDep.getType());
-
-        return new JavaDependency(
-            coordinates,
-            pomDep.getScope(),
-            pomDep.getType(),
-            pomDep.getClassifier(),
-            exclusions
-        );
-    }
-
-    private PartialMavenProject processBoms(String pomFilePath, PartialMavenProject partialModel, Set<String> visitedBoms) throws Exception {
-        int iteration = 0;
-        
-        while (iteration < MAX_BOM_ITERATIONS) {
-            // Build combined properties for resolving BOM coordinates
-            final Map<String, String> currentProps = buildCombinedProperties(partialModel);
-            
-            // Find unprocessed BOMs: scope=import AND not already visited
-            List<PomXmlDependency> unprocessedBoms = partialModel.getDependencyManagement().stream()
-                .filter(dep -> "import".equals(dep.getScope()))
-                .filter(dep -> {
-                    String bomKey = buildBomKey(dep, currentProps);
-                    return !visitedBoms.contains(bomKey);
-                })
-                .collect(Collectors.toList());
-            
-            if (unprocessedBoms.isEmpty()) {
-                logger.debug("No more unprocessed BOMs found after {} iteration(s)", iteration);
-                break;
-            }
-            
-            logger.info("BOM processing iteration {}: found {} unprocessed BOM(s) in {}", 
-                iteration + 1, unprocessedBoms.size(), pomFilePath);
-            
-            // Process each unprocessed BOM
-            for (PomXmlDependency bom : unprocessedBoms) {
-                // Rebuild combined props in case they changed from previous BOM merges
-                final Map<String, String> combinedProps = buildCombinedProperties(partialModel);
-                
-                String resolvedGroupId = resolveProperties(bom.getGroupId(), combinedProps);
-                String resolvedArtifactId = resolveProperties(bom.getArtifactId(), combinedProps);
-                String resolvedVersion = resolveProperties(bom.getVersion(), combinedProps);
-                
-                String bomKey = resolvedGroupId + ":" + resolvedArtifactId + ":" + resolvedVersion;
-                
-                // Mark as visited BEFORE processing to prevent cycles
-                visitedBoms.add(bomKey);
-                logger.debug("Processing BOM: {} (iteration {})", bomKey, iteration + 1);
-                
-                JavaCoordinates bomCoords = new JavaCoordinates(resolvedGroupId, resolvedArtifactId, resolvedVersion, "pom");
-                MavenDownloader mavenDownloader = new MavenDownloader(partialModel.getRepositories(), downloadDir);
-                File bomPomFile = mavenDownloader.downloadPom(bomCoords);
-                
-                if (bomPomFile != null) {
-                    logger.debug("Building BOM project: {}", bomPomFile.getAbsolutePath());
-                    PartialMavenProject bomProject = internalBuildProject(bomPomFile, new HashSet<>());
-                    
-                    // Resolve BOM dependencyManagement entries in BOM context before merging
-                    resolveBomManagedEntries(bomProject);
-                    
-                    // Merge properties from BOM. Existing properties take precedence.
-                    Map<String, String> mergedProperties = new HashMap<>();
-                    if (bomProject.getProperties() != null) {
-                        mergedProperties.putAll(bomProject.getProperties());
-                    }
-                    mergedProperties.putAll(partialModel.getProperties()); // Original properties override BOM's
-                    partialModel.setProperties(mergedProperties);
-                    
-                    // Merge dependency management from BOM. Existing entries take precedence.
-                    // This may introduce NEW scope=import entries (nested BOMs) which will be
-                    // picked up in the next iteration of the while loop.
-                    Map<String, PomXmlDependency> depMgmtMap = new HashMap<>();
-                    if (bomProject.getDependencyManagement() != null) {
-                        bomProject.getDependencyManagement().forEach(dep -> 
-                            depMgmtMap.put(dep.getGroupId() + ":" + dep.getArtifactId(), dep));
-                    }
-                    // Original entries override BOM's entries
-                    partialModel.getDependencyManagement().forEach(dep -> 
-                        depMgmtMap.put(dep.getGroupId() + ":" + dep.getArtifactId(), dep));
-                    partialModel.setDependencyManagement(new ArrayList<>(depMgmtMap.values()));
-                    
-                    logger.debug("Merged BOM {}: {} properties, {} managed dependencies", 
-                        bomKey, 
-                        bomProject.getProperties() != null ? bomProject.getProperties().size() : 0,
-                        bomProject.getDependencyManagement() != null ? bomProject.getDependencyManagement().size() : 0);
-                } else {
-                    logger.warn("Could not download BOM POM for coordinates: {}", bomKey);
-                }
-                
-                // Remove THIS specific processed BOM from the dependency management list
-                // We match by the RESOLVED coordinates to ensure correct removal
-                final String finalGroupId = resolvedGroupId;
-                final String finalArtifactId = resolvedArtifactId;
-                partialModel.getDependencyManagement().removeIf(dep -> 
-                    finalGroupId.equals(resolveProperties(dep.getGroupId(), combinedProps)) &&
-                    finalArtifactId.equals(resolveProperties(dep.getArtifactId(), combinedProps)) &&
-                    "import".equals(dep.getScope()));
-            }
-            
-            iteration++;
-        }
-        
-        if (iteration >= MAX_BOM_ITERATIONS) {
-            logger.warn("Reached maximum BOM processing iterations ({}). Some nested BOMs may not be fully processed. " +
-                "This could indicate a circular BOM import or an unusually deep BOM hierarchy.", MAX_BOM_ITERATIONS);
-        }
-        
-        // Final cleanup: remove any remaining scope=import entries that couldn't be processed
-        int remainingImports = (int) partialModel.getDependencyManagement().stream()
-            .filter(dep -> "import".equals(dep.getScope()))
-            .count();
-        if (remainingImports > 0) {
-            logger.warn("Removing {} unprocessed BOM import(s) from dependency management", remainingImports);
-        }
-        partialModel.getDependencyManagement().removeIf(dep -> "import".equals(dep.getScope()));
-        
-        logger.info("BOM processing complete: processed {} unique BOM(s) in {} iteration(s)", 
-            visitedBoms.size(), iteration);
-        
-        return partialModel;
-    }
-    
-    /**
-     * Builds a combined properties map including parent properties and current model properties.
-     * Used for resolving property placeholders in BOM coordinates.
-     */
-    private Map<String, String> buildCombinedProperties(PartialMavenProject partialModel) {
-        Map<String, String> combinedProps = new HashMap<>();
-        try {
-            Map<String, String> parentProps = propertiesResolverProvider.getParentProperties();
-            if (parentProps != null) {
-                combinedProps.putAll(parentProps);
-            }
-        } catch (Exception e) {
-            logger.debug("Failed to include parent properties while resolving BOM coords: {}", e.getMessage());
-        }
-        if (partialModel.getProperties() != null) {
-            combinedProps.putAll(partialModel.getProperties());
-        }
-        return combinedProps;
-    }
-    
-    /**
-     * Builds a unique key for a BOM dependency using resolved coordinates.
-     * Format: groupId:artifactId:version
-     */
-    private String buildBomKey(PomXmlDependency bom, Map<String, String> properties) {
-        String groupId = resolveProperties(bom.getGroupId(), properties);
-        String artifactId = resolveProperties(bom.getArtifactId(), properties);
-        String version = resolveProperties(bom.getVersion(), properties);
-        return groupId + ":" + artifactId + ":" + version;
-    }
-    
-    /**
-     * Resolves property placeholders in all dependency management entries of a BOM project.
-     * This ensures ${project.version} and similar placeholders become literal values before merging.
-     */
-    private void resolveBomManagedEntries(PartialMavenProject bomProject) {
-        try {
-            Map<String, String> bomProps = new HashMap<>();
-            if (bomProject.getProperties() != null) {
-                bomProps.putAll(bomProject.getProperties());
-            }
-            // Substitute properties for all BOM-managed entries
-            if (bomProject.getDependencyManagement() != null) {
-                for (PomXmlDependency depMgmt : bomProject.getDependencyManagement()) {
-                    depMgmt.setGroupId(resolveProperties(depMgmt.getGroupId(), bomProps));
-                    depMgmt.setArtifactId(resolveProperties(depMgmt.getArtifactId(), bomProps));
-                    depMgmt.setVersion(resolveProperties(depMgmt.getVersion(), bomProps));
-                    depMgmt.setScope(resolveProperties(depMgmt.getScope(), bomProps));
-                    depMgmt.setType(resolveProperties(depMgmt.getType(), bomProps));
-                    depMgmt.setClassifier(resolveProperties(depMgmt.getClassifier(), bomProps));
-                }
-            }
-        } catch (Exception e) {
-            logger.debug("Failed to resolve BOM managed entries in BOM context: {}", e.getMessage());
-        }
-    }
-
-    private PartialMavenProject finalizeEffectiveModel(String path, PartialMavenProject partialModel, PartialMavenProject parentModel) {
-        if (parentModel != null) {
-            logger.info("Merging models: child='{}' into parent='{}'", partialModel.getCoordinates().getArtifactId(), parentModel.getCoordinates().getArtifactId());
-            logger.info("Child properties before merge: {}", partialModel.getProperties().size());
-//            partialModel.getProperties().forEach((k, v) -> logger.info("  - Child Prop: {} = {}", k, v));
-
-            // Merge properties: Child's properties override parent's.
-            Map<String, String> mergedProperties = new HashMap<>();
-            if (parentModel.getProperties() != null) {
-                mergedProperties.putAll(parentModel.getProperties());
-            }
-            if (partialModel.getProperties() != null) {
-                mergedProperties.putAll(partialModel.getProperties());
-            }
-            partialModel.setProperties(mergedProperties);
-            logger.info("Merged properties count: {}", mergedProperties.size());
-//            mergedProperties.forEach((k, v) -> logger.info("  - Merged Prop: {} = {}", k, v));
-
-            // Merge repositories: Child overrides parent by ID
-            Map<String, JavaRepository> repoMap = new HashMap<>();
-            if (parentModel.getRepositories() != null) {
-                parentModel.getRepositories().forEach(repo -> repoMap.put(repo.getId(), repo));
-            }
-            if (partialModel.getRepositories() != null) {
-                partialModel.getRepositories().forEach(repo -> repoMap.put(repo.getId(), repo));
-            }
-            partialModel.setRepositories(new ArrayList<>(repoMap.values()));
-
-            // Merge dependency management: Child overrides parent by G:A
-            Map<String, PomXmlDependency> depMgmtMap = new HashMap<>();
-            if (parentModel.getDependencyManagement() != null) {
-                parentModel.getDependencyManagement().forEach(dep -> depMgmtMap.put(dep.getGroupId() + ":" + dep.getArtifactId(), dep));
-            }
-            if (partialModel.getDependencyManagement() != null) {
-                partialModel.getDependencyManagement().forEach(dep -> depMgmtMap.put(dep.getGroupId() + ":" + dep.getArtifactId(), dep));
-            }
-            partialModel.setDependencyManagement(new ArrayList<>(depMgmtMap.values()));
-
-            // Merge dependencies: Child overrides parent by G:A
-            Map<String, PomXmlDependency> dependenciesMap = new HashMap<>();
-            if (parentModel.getDependencies() != null) {
-                parentModel.getDependencies().forEach(dep -> dependenciesMap.put(dep.getGroupId() + ":" + dep.getArtifactId(), dep));
-            }
-            if (partialModel.getDependencies() != null) {
-                partialModel.getDependencies().forEach(dep -> dependenciesMap.put(dep.getGroupId() + ":" + dep.getArtifactId(), dep));
-            }
-            partialModel.setDependencies(new ArrayList<>(dependenciesMap.values()));
-        }
-        pomCache.put(path, partialModel);
-        return partialModel;
-    }
-
     private MavenProject toCompleteMavenProject(String pomFile, PartialMavenProject project) throws Exception {
         // 1. First pass of property resolution to resolve BOM coordinates
         logger.info("Starting first pass of property resolution for '{}'...", project.getCoordinates().getArtifactId());
-        resolveDependencyProperties(project);
+        dependencyConverter.resolveDependencyProperties(project);
 
         // 2. Process BOMs (including nested BOMs), which may add new properties and managed dependencies.
         // The visitedBoms set tracks processed BOMs to prevent infinite cycles.
         // Process any imported BOMs so their properties and dependencyManagement entries
         // are merged into the partial model before the second property resolution pass.
         Set<String> visitedBoms = new HashSet<>();
-        project = processBoms(pomFile, project, visitedBoms);
+        project = bomProcessor.processBoms(pomFile, project, visitedBoms);
 
         // 3. Second pass of property resolution to handle properties from imported BOMs
         logger.info("Starting second pass of property resolution for '{}' after processing BOMs...", project.getCoordinates().getArtifactId());
-        resolveDependencyProperties(project);
+        dependencyConverter.resolveDependencyProperties(project);
 
         logger.info("Preparing to complete MavenProject. Final properties for resolution in '{}': {}", project.getCoordinates().getArtifactId(), project.getProperties().size());
 //        project.getProperties().forEach((k, v) -> logger.info("  - Final Prop: {} = {}", k, v));
 
         // 4. Apply dependency management to the final set of dependencies
-        List<JavaDependency> finalDependencies = applyDependencyManagement(project);
+        List<JavaDependency> finalDependencies = dependencyConverter.applyDependencyManagement(project);
 
         List<JavaDependency> dependencyManagement = project.getDependencyManagement().stream()
-            .map(this::convertPomXmlDependencyToJavaDependency)
+            .map(dependencyConverter::convertPomXmlDependencyToJavaDependency)
             .collect(Collectors.toList());
 
         return new MavenProject(
@@ -493,9 +194,5 @@ public class ProjectBuilder {
             dependencyManagement,
             project.getModules()
         );
-    }
-
-    private boolean isEmpty(String s) {
-        return s == null || s.trim().isEmpty();
     }
 }
